+++ test/e2e/configurer/chain/commands_btcstaking.go
@@ -28,6 +28,7 @@ import (
 	"github.com/babylonlabs-io/babylon/v4/test/e2e/initialization"
 	"github.com/babylonlabs-io/babylon/v4/testutil/datagen"
 	bbn "github.com/babylonlabs-io/babylon/v4/types"
+	"github.com/babylonlabs-io/babylon/v4/x/btcstaking/types"
 	bstypes "github.com/babylonlabs-io/babylon/v4/x/btcstaking/types"
 )
 
@@ -125,14 +126,86 @@ func (n *NodeConfig) CreateBTCDelegation(
 		cmd = append(cmd, "-b=sync", "--yes")
 	}
 
+	cmd = append(cmd, fmt.Sprintf("--chain-id=%s", n.chainId))
 	outBuff, _, err := n.containerManager.ExecCmd(n.t, n.Name, append(cmd, overallFlags...), "")
 	require.NoError(n.t, err)
 	n.LogActionF("successfully created BTC delegation")
 	return outBuff.String()
 }
 
+func (n *NodeConfig) CreateBTCStakeExpansionDelegation(
+	msg *bstypes.MsgBtcStakeExpand,
+	fromWalletName string,
+	generateOnly bool,
+	overallFlags ...string,
+) (outStr string) {
+	n.LogActionF("creating BTC delegation")
+
+	btcPKHex := msg.BtcPk.MarshalHex()
+
+	// get pop hex
+	popHex, err := msg.Pop.ToHexStr()
+	require.NoError(n.t, err)
+
+	// get staking tx info hex
+	stakingTxHex := hex.EncodeToString(msg.StakingTx)
+
+	fpPKHexList := make([]string, len(msg.FpBtcPkList))
+	for i, fpPK := range msg.FpBtcPkList {
+		fpPKHexList[i] = fpPK.MarshalHex()
+	}
+	fpPKHexes := strings.Join(fpPKHexList, ",")
+
+	stakingTimeString := sdkmath.NewUint(uint64(msg.StakingTime)).String()
+	stakingValueString := sdkmath.NewInt(msg.StakingValue).String()
+
+	// get slashing tx hex
+	slashingTxHex := msg.SlashingTx.ToHexStr()
+	// get delegator sig hex
+	delegatorSigHex := msg.DelegatorSlashingSig.ToHexStr()
+
+	// on-demand unbonding related
+	unbondingTxHex := hex.EncodeToString(msg.UnbondingTx)
+	unbondingSlashingTxHex := msg.UnbondingSlashingTx.ToHexStr()
+	unbondingTimeStr := sdkmath.NewUint(uint64(msg.UnbondingTime)).String()
+	unbondingValueStr := sdkmath.NewInt(msg.UnbondingValue).String()
+	delUnbondingSlashingSigHex := msg.DelegatorUnbondingSlashingSig.ToHexStr()
+
+	fundingTxHex := hex.EncodeToString(msg.FundingTx)
+
+	var inclusionProofHex string
+	cmd := []string{
+		"babylond", "tx", "btcstaking", "btc-stake-expand",
+		btcPKHex, popHex, stakingTxHex, inclusionProofHex, fpPKHexes, stakingTimeString, stakingValueString, slashingTxHex, delegatorSigHex, unbondingTxHex, unbondingSlashingTxHex, unbondingTimeStr, unbondingValueStr, delUnbondingSlashingSigHex,
+		msg.PreviousStakingTxHash, fundingTxHex,
+		fmt.Sprintf("--from=%s", fromWalletName), containers.FlagHome, flagKeyringTest,
+		n.FlagChainID(), "--log_format=json",
+	}
+
+	// gas price
+	cmd = append(cmd, "--gas-prices=0.1ubbn")
+
+	if generateOnly {
+		cmd = append(cmd, "--generate-only")
+	} else {
+		// broadcast stuff
+		cmd = append(cmd, "-b=sync", "--yes")
+	}
+
+	cmd = append(cmd, fmt.Sprintf("--chain-id=%s", n.chainId))
+	outBuff, _, err := n.containerManager.ExecCmd(n.t, n.Name, append(cmd, overallFlags...), "")
+
+	require.NoError(n.t, err)
+	n.LogActionF("successfully created BTC stake expansion delegation")
+	return outBuff.String()
+}
+
 func (n *NodeConfig) AddCovenantSigsFromVal(covPK *bbn.BIP340PubKey, stakingTxHash string, slashingSigs [][]byte, unbondingSig *bbn.BIP340Signature, unbondingSlashingSigs [][]byte) {
-	n.AddCovenantSigs("val", covPK, stakingTxHash, slashingSigs, unbondingSig, unbondingSlashingSigs)
+	n.AddCovenantSigs("val", covPK, stakingTxHash, slashingSigs, unbondingSig, unbondingSlashingSigs, nil)
+}
+
+func (n *NodeConfig) AddCovenantSigsFromValForStakeExp(covPK *bbn.BIP340PubKey, stakingTxHash string, slashingSigs [][]byte, unbondingSig *bbn.BIP340Signature, unbondingSlashingSigs [][]byte, stkExpSig *bbn.BIP340Signature) string {
+	return n.AddCovenantSigs("val", covPK, stakingTxHash, slashingSigs, unbondingSig, unbondingSlashingSigs, stkExpSig)
 }
 
 func (n *NodeConfig) AddCovenantSigs(
@@ -142,7 +215,8 @@ func (n *NodeConfig) AddCovenantSigs(
 	slashingSigs [][]byte,
 	unbondingSig *bbn.BIP340Signature,
 	unbondingSlashingSigs [][]byte,
-) {
+	stakeExpTxSig *bbn.BIP340Signature,
+) string {
 	n.LogActionF("adding covenant signature from nodeName: %s", n.Name)
 
 	covPKHex := covPK.MarshalHex()
@@ -166,14 +240,20 @@ func (n *NodeConfig) AddCovenantSigs(
 	unbondingSlashingSigStr := strings.Join(unbondingSlashingSigStrList, ",")
 	cmd = append(cmd, unbondingSlashingSigStr)
 
+	if stakeExpTxSig != nil {
+		cmd = append(cmd, stakeExpTxSig.ToHexStr())
+	}
+
 	// used key
 	cmd = append(cmd, fmt.Sprintf("--from=%s", fromWalletName))
 	// gas
 	cmd = append(cmd, "--gas-adjustment=2")
 
-	_, _, err := n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
+	outBuf, _, err := n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
 	require.NoError(n.t, err)
 	n.LogActionF("successfully added covenant signatures")
+
+	return GetTxHashFromOutput(outBuf.String())
 }
 
 func (n *NodeConfig) CommitPubRandList(fpBTCPK *bbn.BIP340PubKey, startHeight uint64, numPubrand uint64, commitment []byte, sig *bbn.BIP340Signature) {
@@ -278,7 +358,7 @@ func (n *NodeConfig) BTCUndelegate(
 	spendStakeTx *wire.MsgTx,
 	spendStakeTxInclusionProof *bstypes.InclusionProof,
 	fundingTxs []*wire.MsgTx,
-) {
+) string {
 	n.LogActionF("undelegate by using signature on unbonding tx from delegator")
 
 	spendStakeTxBytes, err := bbn.SerializeBTCTx(spendStakeTx)
@@ -295,11 +375,12 @@ func (n *NodeConfig) BTCUndelegate(
 	}
 	fundingTxsHexStr := strings.Join(fundingTxsHex, ",")
 
-	cmd := []string{"babylond", "tx", "btcstaking", "btc-undelegate", stakingTxHash.String(), spendStakeTxHex, inclusionProofHex, fundingTxsHexStr, "--from=val"}
+	cmd := []string{"babylond", "tx", "btcstaking", "btc-undelegate", stakingTxHash.String(), spendStakeTxHex, inclusionProofHex, fundingTxsHexStr, "--from=val", "--gas=500000"}
 
-	_, _, err = n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
+	outBuf, _, err := n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
 	require.NoError(n.t, err)
 	n.LogActionF("successfully added signature on unbonding tx from delegator")
+	return GetTxHashFromOutput(outBuf.String())
 }
 
 func (n *NodeConfig) AddBTCDelegationInclusionProof(
@@ -472,6 +553,46 @@ func (n *NodeConfig) CreateBTCDel(
 	return testStakingInfo
 }
 
+func (n *NodeConfig) CreateBTCStakeExpDelegationAndCheck(
+	r *rand.Rand,
+	t *testing.T,
+	btcNet *chaincfg.Params,
+	walletNameSender string,
+	fp *bstypes.FinalityProvider,
+	btcStakerSK *btcec.PrivateKey,
+	delAddr string,
+	stakingTimeBlocks uint16,
+	stakingSatAmt int64,
+	prevDel *bstypes.BTCDelegation,
+) (*datagen.TestStakingSlashingInfo, *wire.MsgTx) {
+	msg, testStakingInfo := n.createBtcStakeExpandMessage(
+		r,
+		t,
+		btcNet,
+		btcStakerSK,
+		[]*bstypes.FinalityProvider{fp},
+		stakingSatAmt,
+		stakingTimeBlocks,
+		prevDel,
+	)
+
+	// submit the message for creating BTC delegation
+	n.CreateBTCStakeExpansionDelegation(msg, walletNameSender, false)
+
+	// wait for a block so that above txs take effect
+	n.WaitForNextBlock()
+
+	// check if the address matches
+	btcDelegationResp := n.QueryBtcDelegation(testStakingInfo.StakingTx.TxHash().String())
+	require.NotNil(t, btcDelegationResp)
+	require.Equal(t, btcDelegationResp.BtcDelegation.StakerAddr, delAddr)
+	require.Equal(t, btcStakerSK.PubKey().SerializeCompressed()[1:], btcDelegationResp.BtcDelegation.BtcPk.MustToBTCPK().SerializeCompressed()[1:])
+
+	fundingTx, err := bbn.NewBTCTxFromBytes(msg.FundingTx)
+	require.NoError(t, err)
+	return testStakingInfo, fundingTx
+}
+
 func (n *NodeConfig) AddFinalitySignatureToBlock(
 	fpBTCSK *secp256k1.PrivateKey,
 	fpBTCPK *bbn.BIP340PubKey,
@@ -513,4 +634,348 @@ func CovenantBTCPKs(params *bstypes.Params) []*btcec.PublicKey {
 		covenantBTCPKs[i] = covenantPK.MustToBTCPK()
 	}
 	return covenantBTCPKs
+}
+
+func (n *NodeConfig) createBtcStakeExpandMessage(
+	r *rand.Rand,
+	t *testing.T,
+	btcNet *chaincfg.Params,
+	delSK *btcec.PrivateKey,
+	fps []*bstypes.FinalityProvider,
+	stakingValue int64,
+	stakingTime uint16,
+	prevDel *bstypes.BTCDelegation,
+) (*bstypes.MsgBtcStakeExpand, *datagen.TestStakingSlashingInfo) {
+	// BTC staking params, BTC delegation key pairs and PoP
+	params := n.QueryBTCStakingParams()
+
+	// get fpPKs in BIP340PubKey and BIP340 formats
+	var fpBtcPkList []bbn.BIP340PubKey
+	for _, fp := range fps {
+		fpBtcPkList = append(fpBtcPkList, *fp.BtcPk)
+	}
+	fpPKs, err := bbn.NewBTCPKsFromBIP340PKs(fpBtcPkList)
+	require.NoError(t, err)
+
+	// Convert covenant keys
+	var covenantPks []*btcec.PublicKey
+	for _, pk := range params.CovenantPks {
+		covenantPks = append(covenantPks, pk.MustToBTCPK())
+	}
+
+	// Create funding transaction
+	fundingTx := datagen.GenRandomTxWithOutputValue(r, 10000000)
+
+	// Convert previousStakingTxHash to OutPoint
+	prevDelTxHash := prevDel.MustGetStakingTxHash()
+	prevStakingOutPoint := wire.NewOutPoint(&prevDelTxHash, datagen.StakingOutIdx)
+
+	// Convert fundingTxHash to OutPoint
+	fundingTxHash := fundingTx.TxHash()
+	fundingOutPoint := wire.NewOutPoint(&fundingTxHash, 0)
+	outPoints := []*wire.OutPoint{prevStakingOutPoint, fundingOutPoint}
+
+	// Generate staking slashing info using multiple inputs
+	stakingSlashingInfo := datagen.GenBTCStakingSlashingInfoWithInputs(
+		r,
+		t,
+		btcNet,
+		outPoints,
+		delSK,
+		fpPKs,
+		covenantPks,
+		params.CovenantQuorum,
+		stakingTime,
+		stakingValue,
+		params.SlashingPkScript,
+		params.SlashingRate,
+		uint16(params.UnbondingTimeBlocks),
+	)
+
+	slashingPathSpendInfo, err := stakingSlashingInfo.StakingInfo.SlashingPathSpendInfo()
+	require.NoError(t, err)
+
+	// Sign the slashing tx with delegator key
+	delegatorSig, err := stakingSlashingInfo.SlashingTx.Sign(
+		stakingSlashingInfo.StakingTx,
+		datagen.StakingOutIdx,
+		slashingPathSpendInfo.GetPkScriptPath(),
+		delSK,
+	)
+	require.NoError(t, err)
+
+	// Serialize the staking tx bytes
+	serializedStakingTx, err := bbn.SerializeBTCTx(stakingSlashingInfo.StakingTx)
+	require.NoError(t, err)
+
+	stkTxHash := stakingSlashingInfo.StakingTx.TxHash()
+	unbondingValue := uint64(stakingValue) - uint64(params.UnbondingFeeSat)
+
+	// Generate unbonding slashing info
+	unbondingSlashingInfo := datagen.GenBTCUnbondingSlashingInfo(
+		r,
+		t,
+		btcNet,
+		delSK,
+		fpPKs,
+		covenantPks,
+		params.CovenantQuorum,
+		wire.NewOutPoint(&stkTxHash, datagen.StakingOutIdx),
+		uint16(params.UnbondingTimeBlocks),
+		int64(unbondingValue),
+		params.SlashingPkScript,
+		params.SlashingRate,
+		uint16(params.UnbondingTimeBlocks),
+	)
+
+	unbondingTxBytes, err := bbn.SerializeBTCTx(unbondingSlashingInfo.UnbondingTx)
+	require.NoError(t, err)
+
+	delSlashingTxSig, err := unbondingSlashingInfo.GenDelSlashingTxSig(delSK)
+	require.NoError(t, err)
+
+	// Create proof of possession
+	stakerAddr := sdk.MustAccAddressFromBech32(prevDel.StakerAddr)
+	pop, err := datagen.NewPoPBTC(stakerAddr, delSK)
+	require.NoError(t, err)
+
+	fundingTxBz, err := bbn.SerializeBTCTx(fundingTx)
+	require.NoError(t, err)
+
+	return &types.MsgBtcStakeExpand{
+		StakerAddr:                    prevDel.StakerAddr,
+		Pop:                           pop,
+		BtcPk:                         bbn.NewBIP340PubKeyFromBTCPK(delSK.PubKey()),
+		FpBtcPkList:                   fpBtcPkList,
+		StakingTime:                   uint32(stakingTime),
+		StakingValue:                  stakingValue,
+		StakingTx:                     serializedStakingTx,
+		SlashingTx:                    stakingSlashingInfo.SlashingTx,
+		DelegatorSlashingSig:          delegatorSig,
+		UnbondingValue:                int64(unbondingValue),
+		UnbondingTime:                 params.UnbondingTimeBlocks,
+		UnbondingTx:                   unbondingTxBytes,
+		UnbondingSlashingTx:           unbondingSlashingInfo.SlashingTx,
+		DelegatorUnbondingSlashingSig: delSlashingTxSig,
+		PreviousStakingTxHash:         prevDelTxHash.String(),
+		FundingTx:                     fundingTxBz,
+	}, stakingSlashingInfo
+}
+
+func (n *NodeConfig) SendCovenantSigs(
+	r *rand.Rand,
+	t testing.TB,
+	btcNet *chaincfg.Params,
+	covenantSKs []*btcec.PrivateKey,
+	covWallets []string,
+	pendingDel *bstypes.BTCDelegation,
+) []string {
+	require.Len(t, pendingDel.CovenantSigs, 0)
+
+	params := n.QueryBTCStakingParams()
+	slashingTx := pendingDel.SlashingTx
+	stakingTx := pendingDel.StakingTx
+
+	stakingMsgTx, err := bbn.NewBTCTxFromBytes(stakingTx)
+	require.NoError(t, err)
+	stakingTxHash := stakingMsgTx.TxHash().String()
+
+	fpBTCPKs, err := bbn.NewBTCPKsFromBIP340PKs(pendingDel.FpBtcPkList)
+	require.NoError(t, err)
+
+	stakingInfo, err := pendingDel.GetStakingInfo(params, btcNet)
+	require.NoError(t, err)
+
+	stakingSlashingPathInfo, err := stakingInfo.SlashingPathSpendInfo()
+	require.NoError(t, err)
+
+	/*
+		generate and insert new covenant signature, in order to activate the BTC delegation
+	*/
+	// covenant signatures on slashing tx
+	covenantSlashingSigs, err := datagen.GenCovenantAdaptorSigs(
+		covenantSKs,
+		fpBTCPKs,
+		stakingMsgTx,
+		stakingSlashingPathInfo.GetPkScriptPath(),
+		slashingTx,
+	)
+	require.NoError(t, err)
+
+	// cov Schnorr sigs on unbonding signature
+	unbondingPathInfo, err := stakingInfo.UnbondingPathSpendInfo()
+	require.NoError(t, err)
+	unbondingTx, err := bbn.NewBTCTxFromBytes(pendingDel.BtcUndelegation.UnbondingTx)
+	require.NoError(t, err)
+
+	covUnbondingSigs, err := datagen.GenCovenantUnbondingSigs(
+		covenantSKs,
+		stakingMsgTx,
+		pendingDel.StakingOutputIdx,
+		unbondingPathInfo.GetPkScriptPath(),
+		unbondingTx,
+	)
+	require.NoError(t, err)
+
+	unbondingInfo, err := pendingDel.GetUnbondingInfo(params, btcNet)
+	require.NoError(t, err)
+	unbondingSlashingPathInfo, err := unbondingInfo.SlashingPathSpendInfo()
+	require.NoError(t, err)
+	covenantUnbondingSlashingSigs, err := datagen.GenCovenantAdaptorSigs(
+		covenantSKs,
+		fpBTCPKs,
+		unbondingTx,
+		unbondingSlashingPathInfo.GetPkScriptPath(),
+		pendingDel.BtcUndelegation.SlashingTx,
+	)
+	require.NoError(t, err)
+
+	covStkExpSigs := []*bbn.BIP340Signature{}
+	if pendingDel.IsStakeExpansion() {
+		prevDelTxHash, err := chainhash.NewHash(pendingDel.StkExp.PreviousStakingTxHash)
+		require.NoError(t, err)
+		prevDelRes := n.QueryBtcDelegation(prevDelTxHash.String())
+		require.NotNil(t, prevDelRes)
+		prevDel := prevDelRes.BtcDelegation
+		require.NotNil(t, prevDel)
+		prevParams := n.QueryBTCStakingParamsByVersion(prevDel.ParamsVersion)
+		pDel, err := ParseRespBTCDelToBTCDel(prevDel)
+		require.NoError(t, err)
+		prevDelStakingInfo, err := pDel.GetStakingInfo(prevParams, btcNet)
+		require.NoError(t, err)
+		covStkExpSigs, err = datagen.GenCovenantStakeExpSig(covenantSKs, pendingDel, prevDelStakingInfo)
+		require.NoError(t, err)
+	}
+
+	txHashes := make([]string, params.CovenantQuorum)
+	for i := 0; i < int(params.CovenantQuorum); i++ {
+		// add covenant sigs
+		var stkExpSig *bbn.BIP340Signature
+		if pendingDel.IsStakeExpansion() {
+			stkExpSig = covStkExpSigs[i]
+		}
+		// add covenant sigs
+		txHashes[i] = n.AddCovenantSigs(
+			covWallets[i],
+			covenantSlashingSigs[i].CovPk,
+			stakingTxHash,
+			covenantSlashingSigs[i].AdaptorSigs,
+			bbn.NewBIP340SignatureFromBTCSig(covUnbondingSigs[i]),
+			covenantUnbondingSlashingSigs[i].AdaptorSigs,
+			stkExpSig,
+		)
+		n.WaitForNextBlock()
+	}
+	return txHashes
+}
+
+func (n *NodeConfig) SendCovenantSigsAsValAndCheck(
+	r *rand.Rand,
+	t testing.TB,
+	btcNet *chaincfg.Params,
+	covenantSKs []*btcec.PrivateKey,
+	pendingDel *bstypes.BTCDelegation,
+) {
+	/*
+		generate and insert new covenant signature, in order to activate the BTC delegation
+	*/
+	wallets := make([]string, len(covenantSKs))
+	for i := range covenantSKs {
+		wallets[i] = "val"
+	}
+	txHashes := n.SendCovenantSigs(
+		r, t,
+		btcNet,
+		covenantSKs,
+		wallets,
+		pendingDel,
+	)
+
+	// wait for a block so that above txs take effect
+	n.WaitForNextBlocks(2)
+	for _, txHash := range txHashes {
+		res, _ := n.QueryTx(txHash)
+		require.Equal(t, res.Code, uint32(0), res.RawLog)
+	}
+}
+
+func (n *NodeConfig) CreateBTCDelegationWithExpansionAndCheck(
+	r *rand.Rand,
+	t *testing.T,
+	btcNet *chaincfg.Params,
+	walletNameSender string,
+	fps []*bstypes.FinalityProvider,
+	btcStakerSK *btcec.PrivateKey,
+	delAddr string,
+	stakingTimeBlocks uint16,
+	stakingSatAmt int64,
+	covenantSKs []*btcec.PrivateKey,
+	covenantQuorum uint32,
+) (*datagen.TestStakingSlashingInfo, *datagen.TestStakingSlashingInfo, *wire.MsgTx) {
+	// Step 1: we create a BTC delegation
+	// NOTE: we use the node's address for the BTC delegation
+	prevDelStakingInfo := n.CreateBTCDelegationAndCheck(
+		r,
+		t,
+		btcNet,
+		n.WalletName,
+		fps[0],
+		btcStakerSK,
+		delAddr,
+		stakingTimeBlocks,
+		stakingSatAmt,
+	)
+
+	pendingDelSet := n.QueryFinalityProviderDelegations(fps[0].BtcPk.MarshalHex())
+	require.Len(t, pendingDelSet, 1)
+	pendingDels := pendingDelSet[0]
+	require.Len(t, pendingDels.Dels, 1)
+	require.Equal(t, btcStakerSK.PubKey().SerializeCompressed()[1:], pendingDels.Dels[0].BtcPk.MustToBTCPK().SerializeCompressed()[1:])
+	require.Len(t, pendingDels.Dels[0].CovenantSigs, 0)
+
+	// check delegation
+	delegation := n.QueryBtcDelegation(prevDelStakingInfo.StakingTx.TxHash().String())
+	require.NotNil(t, delegation)
+	require.Equal(t, delegation.BtcDelegation.StakerAddr, n.PublicAddress)
+
+	// Step 2: submit covenant signature to activate the BTC delegation
+	originalDel, err := ParseRespBTCDelToBTCDel(pendingDels.Dels[0])
+	require.NoError(t, err)
+	n.SendCovenantSigsAsValAndCheck(r, t, btcNet, covenantSKs, originalDel)
+
+	// ensure the BTC delegation has covenant sigs now
+	activeDelsSet := n.QueryFinalityProviderDelegations(fps[0].BtcPk.MarshalHex())
+	require.Len(t, activeDelsSet, 1)
+
+	activeDels, err := ParseRespsBTCDelToBTCDel(activeDelsSet[0])
+	require.NoError(t, err)
+	require.NotNil(t, activeDels)
+	require.Len(t, activeDels.Dels, 1)
+
+	activeDel := activeDels.Dels[0]
+	require.True(t, activeDel.HasCovenantQuorums(covenantQuorum, 0))
+
+	// Step 3: create a BTC expansion delegation
+	stkExpDelStakingSlashingInfo, fundingTx := n.CreateBTCStakeExpDelegationAndCheck(
+		r,
+		t,
+		btcNet,
+		n.WalletName,
+		fps[0],
+		btcStakerSK,
+		n.PublicAddress,
+		stakingTimeBlocks,
+		stakingSatAmt,
+		activeDel,
+	)
+
+	// check stake expansion delegation is pending
+	stkExpTxHash := stkExpDelStakingSlashingInfo.StakingTx.TxHash()
+	stkExpDelegation := n.QueryBtcDelegation(stkExpTxHash.String())
+	require.NotNil(t, stkExpDelegation)
+	require.Equal(t, stkExpDelegation.BtcDelegation.StakerAddr, n.PublicAddress)
+	require.NotNil(t, stkExpDelegation.BtcDelegation.StkExp)
+	require.Equal(t, stkExpDelegation.BtcDelegation.StatusDesc, bstypes.BTCDelegationStatus_PENDING.String())
+
+	return stkExpDelStakingSlashingInfo, prevDelStakingInfo, fundingTx
 }