+++ testutil/events/utils.go
@@ -3,6 +3,7 @@ package events
 import (
 	"testing"
 
+	abci "github.com/cometbft/cometbft/abci/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/stretchr/testify/require"
@@ -17,4 +18,8 @@ func RequireEventAttribute(t *testing.T, event sdk.Event, key, expectedValue str
 		}
 	}
 	require.Fail(t, "Expected attribute not found", msgAndArgs...)
+}
+
+func IsEventType(event abci.Event, eventMsg sdk.Msg) bool {
+	return event.Type == sdk.MsgTypeURL(eventMsg) || "/"+event.Type == sdk.MsgTypeURL(eventMsg)
 }